fix user session
