+++ pages/index.js
@@ -3,7 +3,6 @@ import Intro from "@/components/Intro";
 import NavBar from "@/components/NavBar";
 import Head from "next/head";
 import { useDispatch } from "react-redux";
-import { setUserToken, setUserData } from "@/Utils/UserSlice";
 import { useEffect } from "react";
 import Cookies from "js-cookie";
 import { useRouter } from "next/router";
@@ -22,28 +21,7 @@ import { useRouter } from "next/router";
 // }
 
 export default function Home() {
-  const dispatch = useDispatch();
-  const token = Cookies.get("token");
-
-  useEffect(() => {
-    dispatch(
-      setUserData(
-        localStorage.getItem("user")
-          ? JSON.parse(localStorage.getItem("user"))
-          : null
-      )
-    );
-  }, [dispatch]);
-
-  useEffect(() => {
-    if (token) {
-      dispatch(setUserToken(token));
-    } else {
-      localStorage.removeItem("user");
-      dispatch(setUserData(null));
-    }
-  }, [token, dispatch]);
-
+  
 
   return (
     <>
